move to AddWallet namespace
