+++ WalletWasabi.Fluent/ViewModels/AddWallet/RecoverWalletViewModel.cs
@@ -15,7 +15,7 @@ using WalletWasabi.Logging;
 using WalletWasabi.Models;
 using WalletWasabi.Wallets;
 
-namespace WalletWasabi.Fluent.ViewModels.RecoverWallet
+namespace WalletWasabi.Fluent.ViewModels.AddWallet
 {
 	public class RecoverWalletViewModel : RoutableViewModel
 	{
