+++ WalletWasabi.Fluent/ViewModels/AddWalletPageViewModel.cs
@@ -11,7 +11,6 @@ using NBitcoin;
 using WalletWasabi.Fluent.ViewModels.Dialogs;
 using WalletWasabi.Fluent.ViewModels.AddWallet;
 using System.Threading.Tasks;
-using WalletWasabi.Fluent.ViewModels.RecoverWallet;
 using WalletWasabi.Gui.Validation;
 using WalletWasabi.Models;
 using WalletWasabi.Userfacing;
