+++ WalletWasabi.Fluent/Views/AddWallet/RecoverWalletView.axaml
@@ -5,7 +5,7 @@
              xmlns:c="clr-namespace:WalletWasabi.Fluent.Controls"
              xmlns:i="clr-namespace:Avalonia.Xaml.Interactivity;assembly=Avalonia.Xaml.Interactivity"
              xmlns:behaviors="clr-namespace:WalletWasabi.Fluent.Behaviors"
-             xmlns:recoverWallet="clr-namespace:WalletWasabi.Fluent.ViewModels.RecoverWallet"
+             xmlns:recoverWallet="clr-namespace:WalletWasabi.Fluent.ViewModels.AddWallet"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
              x:DataType="recoverWallet:RecoverWalletViewModel"
              x:CompileBindings="True"
