+++ WalletWasabi.Fluent/Views/AddWallet/RecoverWalletView.axaml.cs
@@ -1,7 +1,7 @@
 using Avalonia.Markup.Xaml;
 using Avalonia.ReactiveUI;
 using WalletWasabi.Fluent.ViewModels;
-using WalletWasabi.Fluent.ViewModels.RecoverWallet;
+using WalletWasabi.Fluent.ViewModels.AddWallet;
 
 namespace WalletWasabi.Fluent.Views.RecoverWallet
 {
